Move elements util to root level test/utils
