+++ framework/test/mocha/functional/http/post/3.votes.js
@@ -25,7 +25,7 @@ const phases = require('../../../common/phases');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
-const elements = require('../../../common/utils/elements');
+const elements = require('../../../../utils/elements');
 const apiHelpers = require('../../../common/helpers/api');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
