+++ framework/test/mocha/functional/http/post/7.dapps.out_transfer.js
@@ -22,7 +22,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const phases = require('../../../common/phases');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
-const elements = require('../../../common/utils/elements');
+const elements = require('../../../../utils/elements');
 const apiHelpers = require('../../../common/helpers/api');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
